+++ app/src/main/java/woowacourse/omok/domain/model/game/LobbyManager.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok.domain.model.omokboard
+package woowacourse.omok.domain.model.game
 
 import android.content.Context
 import woowacourse.omok.OmokApplication
