+++ app/src/test/java/woowacourse/omok/domain/model/omokboard/OmokBoardTest.kt
@@ -2,6 +2,7 @@ package woowacourse.omok.domain.model.omokboard
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import woowacourse.omok.domain.model.game.OmokGame
 
 class OmokBoardTest {
     @Test
